# Skip swaths of range tombstone covered keys in merging iterator (2022 edition)