--- a/db/db_iter.cc
+++ b/db/db_iter.cc
@@ -1575,7 +1575,7 @@ void DBIter::SeekToFirst() {
     max_skip_ = std::numeric_limits<uint64_t>::max();
   }
   status_ = Status::OK();
-  // if iterator is empty, this status_ would be unchecked.
+  // if iterator is empty, this status_ could be unchecked.
   status_.PermitUncheckedError();
   direction_ = kForward;
   ReleaseTempPinnedData();
@@ -1646,7 +1646,7 @@ void DBIter::SeekToLast() {
     max_skip_ = std::numeric_limits<uint64_t>::max();
   }
   status_ = Status::OK();
-  // if iterator is empty, this status_ would be unchecked.
+  // if iterator is empty, this status_ could be unchecked.
   status_.PermitUncheckedError();
   direction_ = kReverse;
   ReleaseTempPinnedData();